pyproject.toml
[project]
name = "python-lilypad"
version = "0.0.7"
description = "An open-source prompt engineering framework."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "William Bakst", email = "william@mirascope.io" },
    { name = "Brendan Kao", email = "brendan@mirascope.io" },
]
maintainers = [{ name = "William Bakst", email = "william@mirascope.io" }]
keywords = [
    "prompt engineering",
    "evaluation",
    "llm",
    "llm tools",
    "developer tools",
    "artificial intelligence",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: OpenTelemetry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "mirascope>=1.9.2",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "fastapi[standard]>=0.114.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation>=0.49b1",
    "typer>=0.12.5",
    "workos>=5.7.0",
    "itsdangerous>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.6.1",
]

[project.urls]
Homepage = "https://www.lilypad.so"
Documentation = "https://www.lilypad.so/docs"
Repository = "https://github.com/Mirascope/lilypad"
Issues = "https://github.com/Mirascope/lilypad/issues"
Changelog = "https://github.com/Mirascope/lilypad/releases"

[project.optional-dependencies]
openai = ["openai>=1.43.0"]
anthropic = ["anthropic>=0.36.0"]
gemini = ["google-generativeai>=0.4.0,<1", "pillow>=10.4.0,<11"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.1",
    "pytest>=8.3.2",
    "pyright>=1.1.388",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.8.0",
    "mike>=2.1.3",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.32",
    "mkdocs-macros-plugin>=1.2.0",
    "mkdocstrings>=0.25.2",
    "mkdocstrings-python>=1.10.8",
    "black>=24.8.0",
]

[tool.uv.sources]


[project.scripts]
lilypad = "lilypad.cli.main:app"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
line-length = 88
target-version = "py310"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*.py" = ["S101", "ANN"]
"examples/*.py" = ["T201", "ANN"]
"docs/*.py" = ["ANN401"]

[tool.ruff.lint]
select = [
    "D",
    "E",
    "F",
    "I",
    "UP",
    "YTT",
    "B",
    "T10",
    "T20",
    "C4",
    "PYI006",
    "SIM",
    "ANN",
]
ignore = [
    'D105',
    'D107',
    'D205',
    'D415',
    'E501',
    'B011',
    'B028',
    'B904',
    'ANN101',
    'ANN102',
    'ANN401',
]
fixable = ["ALL"]
isort = { known-first-party = ['lilypad', 'tests', 'examples'] }
pydocstyle = { convention = 'google' }
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
exclude = ["examples"]
venvPath = "."
venv = ".venv"

[tool.codespell]
skip = [".git", "*.lock", "./client/pnpm-lock.yaml"]

[tool.coverage.report]
exclude_also = ["@overload", "@abstractmethod"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lilypad"]

[tool.hatch.build.targets.sdist]
include = ["/lilypad"]

[tool.hatch.build.force-include]
"lilypad/server/static" = "lilypad/server/static"
